--- src/main/java/Entitys/Product.java
@@ -1,16 +1,0 @@
-package Entitys;
-
-public class Product {
-    private int id;
-    Product(int id){
-        this.id = id;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-}
