--- src/main/java/Services/Services.java
@@ -1,11 +1,0 @@
-package Services;
-
-import Entitys.Product;
-
-import java.util.List;
-
-public interface Services {
-    public void addProduct(Product product);
-
-    public List<Product> getProduct();
-}
