+++ src/main/java/com/example/homeworck211/Services/Services.java
@@ -1,0 +1,9 @@
+package com.example.homeworck211.Services;
+
+import java.util.Set;
+
+public interface Services {
+    public void addProduct(int id);
+
+    public Set<Integer> getProduct();
+}
